+++ product_code/certificate_service/api/src/test/scala/de/upb/cs/uc4/certificate/CertificateServiceStub.scala
@@ -2,10 +2,12 @@ package de.upb.cs.uc4.certificate
 
 import akka.{ Done, NotUsed }
 import com.lightbend.lagom.scaladsl.api.ServiceCall
+import com.lightbend.lagom.scaladsl.api.broker.Topic
 import de.upb.cs.uc4.certificate.api.CertificateService
 import de.upb.cs.uc4.certificate.model.{ EncryptedPrivateKey, JsonCertificate, JsonEnrollmentId, PostMessageCSR }
-import de.upb.cs.uc4.shared.client.JsonHyperledgerVersion
+import de.upb.cs.uc4.shared.client.{ JsonHyperledgerVersion, JsonUsername }
 import de.upb.cs.uc4.shared.client.exceptions.UC4Exception
+import de.upb.cs.uc4.shared.client.kafka.EncryptionContainer
 
 import scala.collection.mutable
 import scala.concurrent.Future
@@ -72,6 +74,21 @@ class CertificateServiceStub extends CertificateService {
         case None => Future.failed(UC4Exception.NotFound)
       }
   }
+
+  /** Returns the username that matches the given enrollmentId */
+  override def getUsername(enrollmentId: String): ServiceCall[NotUsed, JsonUsername] = ServiceCall {
+    _ =>
+      Future.successful(JsonUsername(
+        certificateUsers.filter {
+          case (_, entry) => entry.enrollmentId == enrollmentId.trim
+        }.map {
+          case (username, _) => username
+        }.head
+      ))
+  }
+
+  /** Publishes every user that is registered at hyperledger */
+  override def userEnrollmentTopic(): Topic[EncryptionContainer] = null
 
   /** This Methods needs to allow a GET-Method */
   override def allowVersionNumber: ServiceCall[NotUsed, Done] = ServiceCall { _ => Future.successful(Done) }