+++ product_code/hyperledger_component/src/main/scala/de/upb/cs/uc4/hyperledger/commands/HyperledgerProposalCommand.scala
@@ -1,3 +1,5 @@
 package de.upb.cs.uc4.hyperledger.commands
 
-trait HyperledgerProposalCommand extends HyperledgerReadCommand[Array[Byte]]
+import de.upb.cs.uc4.hyperledger.ProposalWrapper
+
+trait HyperledgerProposalCommand extends HyperledgerReadCommand[ProposalWrapper]