+++ product_code/matriculation_service/impl/src/main/scala/de/upb/cs/uc4/matriculation/impl/MatriculationServiceImpl.scala
@@ -11,19 +11,21 @@ import com.typesafe.config.Config
 import de.upb.cs.uc4.authentication.model.AuthenticationRole
 import de.upb.cs.uc4.certificate.api.CertificateService
 import de.upb.cs.uc4.examreg.api.ExamregService
-import de.upb.cs.uc4.hyperledger.HyperledgerUtils
 import de.upb.cs.uc4.hyperledger.commands.{ HyperledgerBaseCommand, SubmitProposal, SubmitTransaction }
+import de.upb.cs.uc4.hyperledger.{ HyperledgerUtils, ProposalWrapper }
 import de.upb.cs.uc4.matriculation.api.MatriculationService
 import de.upb.cs.uc4.matriculation.impl.actor.MatriculationBehaviour
 import de.upb.cs.uc4.matriculation.impl.commands._
 import de.upb.cs.uc4.matriculation.model.{ ImmatriculationData, PutMessageMatriculation }
+import de.upb.cs.uc4.operation.api.OperationService
+import de.upb.cs.uc4.operation.model.JsonOperationId
 import de.upb.cs.uc4.shared.client._
 import de.upb.cs.uc4.shared.client.exceptions._
 import de.upb.cs.uc4.shared.server.ServiceCallFactory._
 import de.upb.cs.uc4.shared.server.messages.{ Accepted, Confirmation, Rejected }
 import de.upb.cs.uc4.user.api.UserService
-import de.upb.cs.uc4.user.model.MatriculationUpdate
 import de.upb.cs.uc4.user.model.user.Student
+import org.slf4j.{ Logger, LoggerFactory }
 import play.api.Environment
 
 import scala.concurrent.duration._
@@ -35,6 +37,7 @@ class MatriculationServiceImpl(
     userService: UserService,
     examregService: ExamregService,
     certificateService: CertificateService,
+    operationService: OperationService,
     override val environment: Environment
 )(implicit ec: ExecutionContext, config: Config, materializer: Materializer)
   extends MatriculationService {
@@ -47,14 +50,12 @@ class MatriculationServiceImpl(
 
   lazy val validationTimeout: FiniteDuration = config.getInt("uc4.timeouts.validation").milliseconds
 
-  /** Submits a proposal to matriculate a student */
-  override def submitMatriculationProposal(username: String): ServiceCall[SignedProposal, UnsignedTransaction] =
-    identifiedAuthenticated[SignedProposal, UnsignedTransaction](AuthenticationRole.Student) { (authUser, _) =>
-      ServerServiceCall { (_, signedProposal) =>
-        if (authUser != username.trim) {
-          throw UC4Exception.OwnerMismatch
-        }
+  private final val log: Logger = LoggerFactory.getLogger(classOf[MatriculationServiceImpl])
 
+  /** Submits a proposal to matriculate a student */
+  override def submitMatriculationProposal(): ServiceCall[SignedProposal, UnsignedTransaction] =
+    authenticated[SignedProposal, UnsignedTransaction](AuthenticationRole.Student, AuthenticationRole.Admin) {
+      ServerServiceCall { (_, signedProposal) =>
         entityRef.askWithStatus[Array[Byte]](replyTo => SubmitProposal(signedProposal, replyTo)).map { array =>
           (ResponseHeader(200, MessageProtocol.empty, List()), UnsignedTransaction(array))
         }.recover(handleException("Submit proposal failed"))
@@ -62,13 +63,9 @@ class MatriculationServiceImpl(
     }
 
   /** Submits a transaction to matriculate a student */
-  override def submitMatriculationTransaction(username: String): ServiceCall[SignedTransaction, Done] =
-    identifiedAuthenticated[SignedTransaction, Done](AuthenticationRole.Student) { (authUser, _) =>
+  override def submitMatriculationTransaction(): ServiceCall[SignedTransaction, Done] =
+    authenticated[SignedTransaction, Done](AuthenticationRole.Student, AuthenticationRole.Admin) {
       ServerServiceCall { (_, signedTransaction) =>
-        if (authUser != username.trim) {
-          throw UC4Exception.OwnerMismatch
-        }
-
         entityRef.askWithStatus[Confirmation](replyTo => SubmitTransaction(signedTransaction, replyTo)).map {
           case Accepted(_) =>
             (ResponseHeader(202, MessageProtocol.empty, List()), Done)
@@ -80,10 +77,10 @@ class MatriculationServiceImpl(
 
   /** Get proposal to matriculate a student */
   override def getMatriculationProposal(username: String): ServiceCall[PutMessageMatriculation, UnsignedProposal] =
-    identifiedAuthenticated[PutMessageMatriculation, UnsignedProposal](AuthenticationRole.Student) { (authUser, _) =>
+    identifiedAuthenticated[PutMessageMatriculation, UnsignedProposal](AuthenticationRole.Student, AuthenticationRole.Admin) { (authUser, authRole) =>
       ServerServiceCall { (header, rawMatriculationProposal) =>
 
-        if (authUser != username.trim) {
+        if (authRole != AuthenticationRole.Admin && authUser != username.trim) {
           throw UC4Exception.OwnerMismatch
         }
 
@@ -122,20 +119,27 @@ class MatriculationServiceImpl(
               .flatMap { jsonEnrollmentId =>
                 val enrollmentId = jsonEnrollmentId.id
 
-                certificateService.getCertificate(username).handleRequestHeader(addAuthenticationHeader(header)).invoke()
+                certificateService.getCertificate(authUser).handleRequestHeader(addAuthenticationHeader(header)).invoke()
                   .flatMap { certificateJson =>
                     val certificate = certificateJson.certificate
 
                     entityRef.askWithStatus[ImmatriculationData](replyTo => GetMatriculationData(enrollmentId, replyTo))
                       .flatMap {
                         _ =>
-                          entityRef.askWithStatus[Array[Byte]](replyTo => GetProposalForAddEntriesToMatriculationData(
+                          entityRef.askWithStatus[ProposalWrapper](replyTo => GetProposalForAddEntriesToMatriculationData(
                             certificate,
                             enrollmentId,
                             matriculationProposal.matriculation,
                             replyTo
                           )).map {
-                            array => (ResponseHeader(200, MessageProtocol.empty, List()), UnsignedProposal(array))
+                            proposalWrapper =>
+                              operationService.addToWatchList(username).handleRequestHeader(addAuthenticationHeader(header)).invoke(JsonOperationId(proposalWrapper.operationId))
+                                .recover {
+                                  case ex: UC4Exception if ex.possibleErrorResponse.`type` == ErrorType.OwnerMismatch =>
+                                  case throwable: Throwable =>
+                                    log.error("Exception in addToWatchlist AddEntries", throwable)
+                                }
+                              (ResponseHeader(200, MessageProtocol.empty, List()), UnsignedProposal(proposalWrapper.proposal))
                           }.recover(handleException("Creation of add entry proposal failed"))
 
                       }.recoverWith {
@@ -144,8 +148,17 @@ class MatriculationServiceImpl(
                             enrollmentId,
                             matriculationProposal.matriculation
                           )
-                          entityRef.askWithStatus[Array[Byte]](replyTo => GetProposalForAddMatriculationData(certificate, data, replyTo)).map {
-                            array => (ResponseHeader(200, MessageProtocol.empty, List()), UnsignedProposal(array))
+                          entityRef.askWithStatus[ProposalWrapper] { replyTo =>
+                            GetProposalForAddMatriculationData(certificate, data, replyTo)
+                          }.map {
+                            proposalWrapper =>
+                              operationService.addToWatchList(username).handleRequestHeader(addAuthenticationHeader(header)).invoke(JsonOperationId(proposalWrapper.operationId))
+                                .recover {
+                                  case ex: UC4Exception if ex.possibleErrorResponse.`type` == ErrorType.OwnerMismatch =>
+                                  case throwable: Throwable =>
+                                    log.error("Exception in addToWatchlist AddMatriculationData", throwable)
+                                }
+                              (ResponseHeader(200, MessageProtocol.empty, List()), UnsignedProposal(proposalWrapper.proposal))
                           }.recover(handleException("Creation of add matriculation data proposal failed"))
 
                         case uc4Exception: UC4Exception => throw uc4Exception
@@ -183,87 +196,11 @@ class MatriculationServiceImpl(
         }
     }
 
-  /** Immatriculates a student */
-  override def addMatriculationData(username: String): ServiceCall[PutMessageMatriculation, Done] =
-    authenticated[PutMessageMatriculation, Done](AuthenticationRole.Admin) {
-      ServerServiceCall { (header, rawMessage) =>
-        val message = rawMessage.trim
-
-        val validationList = try {
-          Await.result(message.validate, validationTimeout)
-        }
-        catch {
-          case _: TimeoutException => throw UC4Exception.ValidationTimeout
-          case e: Exception        => throw UC4Exception.InternalServerError("Validation Error", e.getMessage)
-        }
-
-        if (validationList.nonEmpty) {
-          throw new UC4NonCriticalException(422, DetailedError(ErrorType.Validation, validationList))
-        }
-
-        userService.getUser(username).handleRequestHeader(addAuthenticationHeader(header)).invoke()
-          .flatMap { user =>
-            if (!user.isInstanceOf[Student]) {
-              //We found a user, but it is not a Student. Therefore, a student with the username does not exist: NotFound
-              throw UC4Exception.NotFound
-            }
-
-            certificateService.getEnrollmentId(username).handleRequestHeader(addAuthenticationHeader(header)).invoke()
-              .flatMap { jsonEnrollmentId =>
-                val enrollmentId = jsonEnrollmentId.id
-
-                entityRef.askWithStatus[ImmatriculationData](replyTo => GetMatriculationData(enrollmentId, replyTo))
-                  .flatMap {
-                    _ =>
-                      entityRef.askWithStatus[Confirmation](replyTo => AddEntriesToMatriculationData(
-                        enrollmentId,
-                        message.matriculation,
-                        replyTo
-                      )).map {
-                        case Accepted(_) =>
-                          userService.updateLatestMatriculation().invoke(MatriculationUpdate(
-                            username,
-                            Utils.findLatestSemester(message.matriculation.flatMap(_.semesters))
-                          ))
-                          (ResponseHeader(200, MessageProtocol.empty, List()), Done)
-
-                        case Rejected(statusCode, reason) => throw UC4Exception(statusCode, reason)
-                      }.recover(handleException("Error in AddEntriesToMatriculationData"))
-                  }.recoverWith {
-                    case uc4Exception: UC4Exception if uc4Exception.errorCode == 404 =>
-                      val data = ImmatriculationData(
-                        enrollmentId,
-                        message.matriculation
-                      )
-                      entityRef.askWithStatus[Confirmation](replyTo => AddMatriculationData(data, replyTo)).map {
-                        case Accepted(_) =>
-                          userService.updateLatestMatriculation().invoke(MatriculationUpdate(
-                            username,
-                            Utils.findLatestSemester(message.matriculation.flatMap(_.semesters))
-                          ))
-                          (ResponseHeader(201, MessageProtocol.empty, List(("Location", s"$pathPrefix/history/$username"))), Done)
-
-                        case Rejected(statusCode, reason) => throw UC4Exception(statusCode, reason)
-                      }.recover(handleException("Error in AddMatriculationData"))
-
-                    case uc4Exception: UC4Exception => throw uc4Exception
-
-                    case ex: Throwable =>
-                      throw UC4Exception.InternalServerError("Failure at addition of new matriculation data", ex.getMessage, ex)
-                  }
-              }
-          }
-      }
-    }
-
   /** Allows GET */
   override def allowedGet: ServiceCall[NotUsed, Done] = allowedMethodsCustom("GET")
 
   /** Allows POST */
   override def allowedPost: ServiceCall[NotUsed, Done] = allowedMethodsCustom("POST")
-
-  /** Allows PUT */
-  override def allowedPut: ServiceCall[NotUsed, Done] = allowedMethodsCustom("PUT")
 
   /** This Methods needs to allow a GET-Method */
   override def allowVersionNumber: ServiceCall[NotUsed, Done] = allowedMethodsCustom("GET")