+++ product_code/shared/client/src/main/scala/de/upb/cs/uc4/shared/client/operation/OperationDataState.scala
@@ -1,0 +1,12 @@
+package de.upb.cs.uc4.shared.client.operation
+
+import play.api.libs.json.{ Format, Json }
+
+object OperationDataState extends Enumeration {
+  type OperationDataState = Value
+  val PENDING, FINISHED, REJECTED = Value
+
+  implicit val format: Format[OperationDataState] = Json.formatEnum(this)
+
+  def All: Seq[OperationDataState] = values.toSeq
+}