+++ product_code/shared/client/src/main/scala/de/upb/cs/uc4/shared/client/operation/TransactionInfo.scala
@@ -1,0 +1,9 @@
+package de.upb.cs.uc4.shared.client.operation
+
+import play.api.libs.json.{ Format, Json }
+
+case class TransactionInfo(contractName: String, transactionName: String, parameters: String)
+
+object TransactionInfo {
+  implicit val format: Format[TransactionInfo] = Json.format
+}