+++ product_code/shared/server/src/test/scala/de/upb/cs/uc4/shared/server/KafkaEncryptionUtilitySpec.scala
@@ -1,11 +1,10 @@
 package de.upb.cs.uc4.shared.server
 
 import java.security.InvalidKeyException
-
-import de.upb.cs.uc4.authentication.model.JsonUsername
-import de.upb.cs.uc4.shared.client.JsonServiceVersion
+import de.upb.cs.uc4.shared.client.{ JsonServiceVersion, JsonUsername }
 import de.upb.cs.uc4.shared.client.exceptions.{ ErrorType, UC4Exception }
 import de.upb.cs.uc4.shared.server.kafka.KafkaEncryptionUtility
+
 import javax.crypto.AEADBadTagException
 import org.scalatest.PrivateMethodTester
 import org.scalatest.compatible.Assertion